divided property page into 5 sections, and fixed dimensions and button of the calander
